Make the bf command fade smoothly, instead of rounding down and disappearing almost instantly.

+++ quakespasm/Quake/client.h
@@ -46,7 +46,7 @@ typedef struct
 typedef struct
 {
 	int		destcolor[3];
-	int		percent;		// 0-256
+	float	percent;		// 0-256
 } cshift_t;
 
 #define	CSHIFT_CONTENTS	0
